agent/requirements.txt
livekit-agents[codecs,images,turn-detector]~=1.4.3
livekit-plugins-deepgram~=1.0
livekit-plugins-openai~=1.0
livekit-plugins-silero~=1.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0
edge-tts>=6.1
